// Client/src/app/Models/TravelDetail.ts
export class TravelDetail {
    TravelDetailId: number;
    RequisitionId: number;
    VehicleId: number;
    DriverId: number;
    MileageStart: number;
    MileageEnd: number;
    FuelInLiter: number;
    Amount: number;
    ReceiptNo: string;
    FilledBy: string;
    TravelStartTime: string;
    TravelEndTime: string;
    VisitingPlace: string;
}